refactor llama2.py
